pyproject.toml
[project]
name = "gabos-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Gabriel Torres Gamez", email = "gabriel.torres@fuet.ch" }
]
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.13.0",
    "fastmcp>=3.0.0",
    "html2text>=2024.2.26",
    "platformdirs>=4.9.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.2",
    "ty>=0.0.18",
    "whoosh>=2.7.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
gabos-mcp = "gabos_mcp:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
extend-select = [
    "W",   # pycodestyle warnings
    "I",   # import sorting
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "ISC", # implicit string concat
    "RET", # return practices
    "SIM", # simplification
    "TID", # import practices
    "TC",  # TYPE_CHECKING block enforcement
    "PTH", # pathlib over os.path
    "D",   # pydocstyle
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]  # no docstring requirements for tests

[tool.ruff.lint.pydocstyle]
convention = "google"
